+++ packages/app/package.json
@@ -46,6 +46,7 @@
     "@backstage/theme": "0.4.1",
     "@janus-idp/backstage-plugin-acr": "1.1.0",
     "@janus-idp/backstage-plugin-jfrog-artifactory": "1.1.0",
+    "@janus-idp/backstage-plugin-ocm": "3.1.0",
     "@janus-idp/backstage-plugin-quay": "1.3.0",
     "@janus-idp/backstage-plugin-tekton": "1.9.0",
     "@janus-idp/backstage-plugin-topology": "1.15.0",
