+++ packages/app/src/components/catalog/EntityPage/Content/Overview.tsx
@@ -1,0 +1,40 @@
+import {
+  EntityAboutCard,
+  EntityHasSubcomponentsCard,
+  EntityLinksCard,
+  EntitySwitch,
+} from '@backstage/plugin-catalog';
+import { EntityCatalogGraphCard } from '@backstage/plugin-catalog-graph';
+import React from 'react';
+import Grid from '@mui/material/Grid';
+import {
+    EntityArgoCDHistoryCard,
+    isArgocdAvailable,
+  } from '@roadiehq/backstage-plugin-argo-cd';
+import { entityWarningContent } from './EntityWarning';
+
+export const overviewContent = (
+  <Grid container spacing={3} alignItems="stretch">
+    {entityWarningContent}
+    <Grid item md={6}>
+      <EntityAboutCard variant="gridItem" />
+    </Grid>
+    <Grid item md={6} xs={12}>
+      <EntityCatalogGraphCard variant="gridItem" height={400} />
+    </Grid>
+
+    <Grid item md={4} xs={12}>
+      <EntityLinksCard />
+    </Grid>
+    <Grid item md={8} xs={12}>
+      <EntityHasSubcomponentsCard variant="gridItem" />
+    </Grid>
+    <EntitySwitch>
+      <EntitySwitch.Case if={isArgocdAvailable}>
+        <Grid item sm={6}>
+          <EntityArgoCDHistoryCard />
+        </Grid>
+      </EntitySwitch.Case>
+    </EntitySwitch>
+  </Grid>
+);
