+++ packages/app/src/components/catalog/EntityPage/Content/index.tsx
@@ -1,0 +1,4 @@
+import * from './CI-CD'
+import * from 'EntityWarning'
+import * from './Overview'
+import * from './Techdocs'
