+++ packages/app/src/components/catalog/EntityPage/Pages/API.tsx
@@ -1,0 +1,49 @@
+import {
+  EntityProvidingComponentsCard,
+  EntityConsumingComponentsCard,
+  EntityApiDefinitionCard,
+} from '@backstage/plugin-api-docs';
+import {
+  EntityLayout,
+  EntityAboutCard,
+  EntityLinksCard,
+} from '@backstage/plugin-catalog';
+import { EntityCatalogGraphCard } from '@backstage/plugin-catalog-graph';
+import { Grid } from '@material-ui/core';
+import { entityWarningContent } from '../Content/EntityWarning';
+import React from 'react';
+
+export const apiPage = (
+  <EntityLayout>
+    <EntityLayout.Route path="/" title="Overview">
+      <Grid container spacing={3}>
+        {entityWarningContent}
+        <Grid item md={6}>
+          <EntityAboutCard />
+        </Grid>
+        <Grid item md={6} xs={12}>
+          <EntityCatalogGraphCard variant="gridItem" height={400} />
+        </Grid>
+        <Grid item md={4} xs={12}>
+          <EntityLinksCard />
+        </Grid>
+        <Grid container item md={12}>
+          <Grid item md={6}>
+            <EntityProvidingComponentsCard />
+          </Grid>
+          <Grid item md={6}>
+            <EntityConsumingComponentsCard />
+          </Grid>
+        </Grid>
+      </Grid>
+    </EntityLayout.Route>
+
+    <EntityLayout.Route path="/definition" title="Definition">
+      <Grid container spacing={3}>
+        <Grid item xs={12}>
+          <EntityApiDefinitionCard />
+        </Grid>
+      </Grid>
+    </EntityLayout.Route>
+  </EntityLayout>
+);
