+++ packages/app/src/components/catalog/EntityPage/Pages/index.tsx
@@ -1,0 +1,8 @@
+export * from './API';
+export * from './Component';
+export * from './DefaultEntity';
+export * from './Domain';
+export * from './Group';
+export * from './ResourcePage';
+export * from './System';
+export * from './User';
