+++ packages/backend/package.json
@@ -36,6 +36,7 @@
     "@backstage/plugin-techdocs-backend": "1.6.4",
     "@janus-idp/backstage-plugin-aap-backend": "1.1.0",
     "@janus-idp/backstage-plugin-keycloak-backend": "1.5.0",
+    "@janus-idp/backstage-plugin-ocm-backend": "3.1.0",
     "@roadiehq/backstage-plugin-argo-cd-backend": "2.11.0",
     "@types/react": "^17",
     "@types/react-dom": "^17",
